@@ -3,18 +3,18 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "node-sass": "^4.13.1",
-    "react": "^16.12.0",
-    "react-accessible-accordion": "^3.0.1",
-    "react-dom": "^16.12.0",
-    "react-icons": "^3.9.0",
-    "react-parallax": "^3.0.3",
-    "react-player": "^1.14.2",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.3.0"
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.4.3",
+    "node-sass": "^7.0.1",
+    "react": "^18.2.0",
+    "react-accessible-accordion": "^5.0.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.4.0",
+    "react-parallax": "^3.5.1",
+    "react-player": "^2.10.1",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
